server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.34.0
supabase==2.13.0
python-dotenv==1.1.0
pydantic[email]==2.11.1
